+++ CHANGES.rst
@@ -8,6 +8,7 @@ Changelog
 Added:
 
 - Added Clean and Demo profiles with demo content. #6
+- Added configurable section colors to configlet. #5
   [netroxen]
 
 - Initial release.